--- static/css/styles.css
@@ -1,3 +1,0 @@
-nav a{
-  font-weight: bold;
-}
