+++ vscribe/__init__.py
@@ -1,0 +1,27 @@
+from flask import Flask
+import pymongo
+from logging.config import dictConfig
+
+dictConfig({
+    'version': 1,
+    'formatters': {'default': {
+        'format': '[%(asctime)s] %(levelname)s in %(module)s: %(message)s',
+    }},
+    'handlers': {'wsgi': {
+        'class': 'logging.StreamHandler',
+        'stream': 'ext://flask.logging.wsgi_errors_stream',
+        'formatter': 'default'
+    }},
+    'root': {
+        'level': 'INFO',
+        'handlers': ['wsgi']
+    }
+})
+
+app = Flask(__name__)
+
+myclient = pymongo.MongoClient("mongodb://localhost:27017/")
+mydb = myclient["vscribe"]
+mycol = mydb["qna"]
+
+from vscribe import routes
