+++ vscribe/static/css/styles.css
@@ -1,0 +1,3 @@
+nav a{
+  font-weight: bold;
+}
