+++ vscribe/templates/test.htm
@@ -1,0 +1,76 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <style media="screen">
+      body{
+        height: 1000px;
+      }
+      nav a{
+        font-weight: bold;
+        margin: 10px 20px 10px 20px;
+      }
+      .home img{
+        width: 100%;
+        height: 70%;
+        opacity: 0.6;
+      }
+      .content{
+        position: absolute;
+        top: 100px;
+        margin: auto;
+        text-align: center;
+        left: 500px;
+      }
+      .dt{
+        margin: auto;
+        margin-top: 20px;
+        text-align: center;
+      }
+      h1{
+        font-family: 'Montserrat', sans-serif;
+      }
+      #answer{
+        margin-bottom: 30px;
+        -moz-box-shadow: 0 0 5px #888;
+        -webkit-box-shadow: 0 0 5px#888;
+        box-shadow: 0 0 5px #888;
+      }
+      #qno{
+        background-color: black;
+        border-radius: 5px;
+        padding: 2px 5px 2px 5px;
+        color: white;
+        margin-right: 5px;
+      }
+    </style>
+      <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+    <script src="{{url_for('static',filename='process.js')}}"></script>
+    <link href="https://fonts.googleapis.com/css?family=Montserrat:400,500&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css"
+        integrity="sha384-50oBUHEmvpQ+1lW4y57PTFmhCaXp0ML5d60M1M7uH2+nqUivzIebhndOJK28anvf" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
+    <meta charset="utf-8">
+    <title>Questions</title>
+
+
+
+  </head>
+  <body>
+    <nav class="navbar navbar-dark bg-dark">
+  <a class="navbar-brand" href="#"><i class="fas fa-pencil-ruler"></i> V-Scribe</a>
+    <ul class="navbar-nav">
+      <li class="nav-item active">
+        <a class="nav-link" href="#"><i class="fas fa-smile"></i> Finished</a>
+      </li>
+    </ul>
+</nav>
+    <div class="dt" style="margin-top:140px">
+        <form action="/brail" method="GET" style="margin: 20px auto 20px auto; width: 100px">
+          <button type="submit" name="button" class="btn btn-secondary"><i class="fas fa-download"></i> Braille</button>
+        </form>
+      <h1>Thanks for writing the paper. All the best for results.</h1>
+    </div>
+  </body>
+</html>
